requirements.txt
absl-py==0.12.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cv2-tools==2.4.0
dill==0.3.3
EasyProcess==0.3
fer==0.0.1
Flask==1.1.2
future==0.18.2
googleapis-common-protos==1.53.0
idna==2.10
ImageHash==4.2.0
importlib-resources==5.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.1
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.1.2
promise==2.3
protobuf==3.15.6
pyparsing==2.4.7
PyQt5==5.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
PyQtWebKit==5.13.1
python-constraint==1.4.0
PyVirtualDisplay==2.1
PyWavelets==1.1.1
Qt.py==1.3.3
requests==2.25.1
scipy==1.6.2
selenium==3.141.0
sip==4.19.8
six==1.15.0
tensorflow-datasets==4.2.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
